requirements.txt
asgiref==3.8.1
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
Django==5.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
idna==3.10
inflection==0.5.1
lxml==5.4.0
markdown-it-py==3.0.0
mdurl==0.1.2
packaging==25.0
Pygments==2.19.1
PyJWT==2.9.0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
rich==14.0.0
sqlparse==0.5.3
uritemplate==4.1.1
urllib3==2.4.0
